requirements.txt
langchain
langchain-google-genai
pillow
pydantic
python-dotenv
google-generativeai
gspread
oauth2client
